add ( step_2 ) : add widgets
